Update pet api endpoints
