+++ pages/holding/[id].js
@@ -1,9 +1,9 @@
-import Adoption from "../../components/Adoption";
+import HoldingScreen from "../../components/HoldingScreen";
 import React from "react";
 import axios from "axios";
 
-const Build = (props) => {
-	return <Adoption {...props} />;
+const Holding = (props) => {
+	return <HoldingScreen {...props} />;
 };
 
 export async function getServerSideProps({ params }) {
@@ -17,4 +17,4 @@ export async function getServerSideProps({ params }) {
 	};
 }
 
-export default Build;
+export default Holding;
